Added ssh_agent function
ssh_agent finds an existing ssh-agent and configures environment
variables to use it. If it cannot find an existing instance, it starts a
new one.

The zshexit() hook will check to see if there are any remaining zsh
instances on the current machine, and will kill ssh-agent if the exiting
shell is the last zsh instance.