Added OS detection to ssh_agent
OS X handles ssh-agent using Keychain, so it's not necessary to detect
existing instances of ssh-agent.

However, OS X also uses /var/folders instead of /tmp, which is why this was
even an issue in the first place. Added SSH_TMP_PATH variable and case
statement to deal with situations where the OS is not supported or where /tmp
is not the correct path to search.